% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_allocation_parameters.r
\name{calibrate_allocation_parameters}
\alias{calibrate_allocation_parameters}
\title{Calculate_allocation_parameters:}
\usage{
calibrate_allocation_parameters(config = get_config())
}
\arguments{
\item{config}{a list of configuration params}
}
\description{
Using land use data from calibration (historic) periods to calculate parameters for
Dinamica's Patcher and Expander algorithmns Mean patch size, patch size variance,
patch isometry, percentage of transitons in new patches vs. expansion of existing
patches
}
\details{
\enumerate{
\item Estimate values for the allocation parameters and then apply random perturbation
to generate sets of values to test with monte-carlo simulation
\item perform simulations with all parameter sets
\item Identify best performing parameter sets and save copies of tables
to be used in scenario simulations
}
}
\author{
Ben Black
}
