% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_predictor_prep.r
\name{calibration_predictor_prep}
\alias{calibration_predictor_prep}
\title{Preparation of Predictor Data}
\usage{
calibration_predictor_prep(
  config = get_config(),
  refresh_cache = FALSE,
  ignore_excel = FALSE
)
}
\arguments{
\item{config}{A list object}
}
\description{
Does some predictor data preparation, relying on the xlsx file at \code{pred_table_path}
and other assumptions specified in \code{\link[=get_config]{get_config()}}.
}
