% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_initialize.r
\name{create_init_lulc_raster}
\alias{create_init_lulc_raster}
\title{Creates the initial LULC raster and copies it into the results directory}
\usage{
create_init_lulc_raster(params = get_simulation_params())
}
\description{
Creates the initial LULC raster and copies it into the results directory
}
