% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_initialize.r
\name{default_ctrl_tbl_path}
\alias{default_ctrl_tbl_path}
\title{Get Simulation Parameters}
\usage{
default_ctrl_tbl_path()
}
\description{
This functionality is roughly equivalent to what used to be called
\code{Dinamica_initialize.r}. It is intended to be called from within a Dinamica
simulation to read parameters.
Date: 25-02-2022
Author: Ben Black
}
