% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_trans_prep.r
\name{deterministic_trans_prep}
\alias{deterministic_trans_prep}
\title{Process deterministic transitions (glacier meltoff)}
\usage{
deterministic_trans_prep(config = get_config())
}
\arguments{
\item{config}{A configuration list, default is retrieved from get_config()}
}
\value{
No direct return value. The function saves:
\itemize{
\item RDS files of glacial change indices for each scenario
\item An Excel file with glacial area change data across scenarios
}
}
\description{
Prepares data for deterministic transitions in the land use change model. Currently
only entails reading pre-treated data for where-on-the-grid glaciers will disappear.
}
\details{
This function performs two main tasks:
\enumerate{
\item Preparation of general model data:
\itemize{
\item Extracts years from historic LULC files
\item Loads transition model lookup tables
\item Prepares time step vectors for simulation, which incidentally are the same that
the source data have.
}
\item Processing of glacial change data:
\itemize{
\item Reads glacial location data from the Farinotti et al. dataset
\item Calculates glacial coverage changes for each time step and scenario
\item Saves RCP-specific glacial indices for later use
\item Creates a summary table of glacial areal changes across scenarios
}
}
}
\author{
Ben Black
}
