% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_deterministic_trans.r
\name{dinamica_deterministic_trans}
\alias{dinamica_deterministic_trans}
\title{dinamica_deterministic_trans}
\usage{
dinamica_deterministic_trans(
  simulated_lulc_save_path,
  simulation_id,
  simulated_lulc_year,
  config = get_config()
)
}
\arguments{
\item{simulated_lulc_save_path}{Path (char) passed in from Dinamica}

\item{simulation_id}{integerish}

\item{simulated_lulc_year}{integerish}

\item{config}{list of config options}
}
\description{
Implement deterministic transitions following statistical transition potential during
during simulation step
}
\author{
Ben Black
}
