% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_load_trans_table.r
\name{dinamica_load_trans_table}
\alias{dinamica_load_trans_table}
\title{dinamica_load_trans_table}
\usage{
dinamica_load_trans_table(simulation_id, year, config = get_config())
}
\arguments{
\item{simulation_id}{Integer. The ID of the simulation to load parameters for.}

\item{year}{Integer. The year for which to load the transition matrix.}

\item{config}{List. Configuration settings, defaults to the result of \code{\link[=get_config]{get_config()}}.}
}
\value{
A data frame containing the transition table ("From*" (int), "To*" (int),
"Rate" (float))
}
\description{
Loads a scenario and year-specific transition matrix for a simulation, and updates it
to reflect any deterministic transitions (such as glacier transitions).
}
\author{
Ben Black
}
