% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_update_control_table.r
\name{dinamica_update_control_table}
\alias{dinamica_update_control_table}
\title{Update the control table after completion of simulation}
\usage{
dinamica_update_control_table(
  control_table_path = character(),
  success = character(),
  simulation_num = integer()
)
}
\arguments{
\item{control_table_path}{Character. Path to the control table CSV file.}

\item{success}{Character. Indicates if the simulation was successful ("TRUE") or not.}

\item{simulation_num}{Integer. The simulation number to update in the control table.}
}
\value{
None. The function is called for its side effect of updating the CSV file.
}
\description{
This function updates the "completed" status in a control table CSV file after a simulation run.
}
\details{
The function reads the control table, updates the "completed.string" column for the
specified simulation number to "Y" if the simulation was successful, or "ERROR"
otherwise, and writes the updated table back to disk.
}
\author{
Ben Black
}
