% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_use_validation.r
\name{dinamica_use_validation}
\alias{dinamica_use_validation}
\title{dinamica_use_validation}
\usage{
dinamica_use_validation(
  config = get_config(),
  simulation_id = integer(),
  sim_results_path
)
}
\arguments{
\item{simulation_id}{integerish}

\item{sim_results_path}{character}
}
\value{
A list of \verb{validation_condition, Validation_result_path, Validation_map_path, Final_LULC_path, Sim_final_LULC_path}
}
\description{
Determine if validation should be used, and what paths to use for validation. If run
in simulation mode (i.e. control table \code{model_mode.string} matches "simulation") then
no validation will take place.
}
