% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{makes datatable-aware, so NSE works}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
makes datatable-aware, so NSE works
}
\keyword{datasets}
