% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{get_config}
\alias{get_config}
\title{LULCC_CH setup: control table, config}
\usage{
get_config(
  scenario_names = c("BAU", "EI-NAT", "EI-CUL", "EI-SOC", "GR-EX"),
  step_length = 5L,
  data_periods = c("1985_1997", "1997_2009", "2009_2018"),
  regionalization = TRUE,
  inclusion_threshold = 0.5
)
}
\arguments{
\item{scenario_names}{character vector of scenario names}

\item{step_length}{integer step length for simulation}

\item{data_periods}{character vector of data periods}

\item{regionalization}{logical whether to regionalize}

\item{inclusion_threshold}{numeric, minimum \% threshold for transition inclusion.
This represents the number of transition instances from class X -> Y as a \% of the
the total area of class X; a good value for this threshold is 0.5 such that if the
number of cells transitioning <0.5\% of the total number of cells of the initial class
then the transition is not included. The rationale for this is that the statistical
model produced for the transition will be too weak due to high-imbalance}

\item{simctrl_tbl_path}{character path to simulation control table}
}
\value{
list of configuration parameters
}
\description{
Creates / reads control table and config files
}
