% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.analysecovselection.r
\name{lulcc.analysecovselection}
\alias{lulcc.analysecovselection}
\title{lulcc.analysecovselection}
\usage{
lulcc.analysecovselection(All_pred_names, Filtered_predictors, summary_level)
}
\arguments{
\item{All_pred_names}{list of unique covariate names
produced by function lulcc.evalfeatureselection}

\item{Filtered_predictors}{list object containing results of
covariate selection produced by function lulcc.evalfeatureselection}

\item{summary_level}{Characters describing aspect of transition
datasets to summarize by accepts:
"across_trans", "Bioregion", "Initial_lulc", "Final_lulc".}
}
\description{
Produce summary information on results of 2 stage covariate selection procedure
}
\author{
Ben Black
}
