% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.downloadunzip.r
\name{lulcc.downloadunzip}
\alias{lulcc.downloadunzip}
\title{Download and unzip a file}
\usage{
lulcc.downloadunzip(
  url = character(),
  save_dir = character(),
  filename = basename(url),
  force_lowercase = FALSE
)
}
\arguments{
\item{url}{character string url to download folder from
produced by function lulcc.evalfeatureselection}

\item{filename}{defaults to the \code{basename(url)}}

\item{save_path}{character string path of directory to save unzipped files in}
}
\description{
Download folder from URL and unzip to target directory
}
\author{
Ben Black
}
