% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.evalcurves.r
\name{lulcc.evalcurves}
\alias{lulcc.evalcurves}
\title{Evaluate models}
\usage{
lulcc.evalcurves(
  x,
  tester = data.frame(),
  thres = numeric(),
  crit = "pp=op",
  prevalence_correction = FALSE
)
}
\arguments{
\item{x}{a wsl.fit object}

\item{tester}{data.frame with testing data (only mandatory if replicatetype='none'
was chosen when models were fitted)}

\item{crit}{which threshold criterion should be considered? Currently 'pp=op'
(predicted prevalence = observed prevalence), 'maxTSS' (threshold yielding maximum TSS),
and 'external' (thresholds manually supplied) are possible}

\item{prevalence_correction}{logical. Should imbalanced presence/absence data be upsampled to
prevalence 0.5 for model evaluation.
data. In a pseudoabsence correction the same amount of pseudoabsences as presences are
subsampled for model validation.}

\item{threshold}{vector of the same length as number of models chosen with custom
thresholds for model evaluation. for wsl.flex outputs the thresholds have to be labelled
with the same names provided to models}
}
\value{
an obejct of class 'wsl.evaluation'
}
\description{
Produce objects needed for Boyce and ROC curves of all models in a wsl.fit object.
}
\author{
Philipp Brun (main) with edits by Antoine Adde and Ben Black
}
