% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.evalfeatureselection2.r
\name{lulcc.evalfeatureselection2}
\alias{lulcc.evalfeatureselection2}
\title{lulcc.evalfeatureselection}
\usage{
lulcc.evalfeatureselection2(
  Predictor_table_path,
  data_period_name,
  Dataset_scale,
  Pre_FS_folder,
  FS_results_folder,
  summary_levels
)
}
\arguments{
\item{data_period_name}{Character name of dataset period to be used in file saving.}

\item{Dataset_scale}{Character, scale of datasets to be used in file saving.}

\item{Pre_FS_folder}{Character, folder path for pre-covariate selection datasets.}

\item{FS_results_folder}{Character folder path to save summaries.}

\item{summary_levels}{vector of characters describing aspect of transition
datasets to summarize by accepts:
"across_trans", "Bioregion", "Initial_lulc", "Final_lulc".}
}
\description{
Wrapper function to summarise the results of the feature selection process according
to different aspects of the transitions datasets
}
\author{
Ben Black
}
