% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.filtersel.r
\name{lulcc.filtersel}
\alias{lulcc.filtersel}
\title{lulcc.filtersel}
\usage{
lulcc.filtersel(
  transition_result,
  cov_data,
  categories,
  collin_weight_vector,
  embedded_weight_vector,
  focals = NULL,
  method,
  corcut = 0
)
}
\arguments{
\item{transition_result}{a vector containing binary values of transition result (0/1)}

\item{cov_data}{a datatable containing column wise values of covariates}

\item{categories}{a vector of strings of covariate categories}

\item{collin_weight_vector}{vector of weights used for collinearity filtering}

\item{embedded_weight_vector}{weights used for embedded (grrf) filtering}

\item{focals}{a character vector identifying focal covariates}

\item{method}{a character of the method used to rank covariates
accepts: "IQR", "IQR.M", "COR.P", "COR.S", "WIL", "GLM", "MAD.M"}

\item{corcut}{numeric value for correlation cut-off to remove covariates}
}
\description{
Apply the collinearity filtering on groups of predictors
}
\author{
Antoine Adde (main) edited by Ben Black
}
