% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.fitevalsave.r
\name{lulcc.fitevalsave}
\alias{lulcc.fitevalsave}
\title{lulcc.fitevalsave}
\usage{
lulcc.fitevalsave(
  Transition_dataset,
  Transition_name,
  replicatetype,
  reps,
  balance_class,
  Downsampling_bounds,
  Data_period,
  model_folder,
  eval_results_folder,
  Model_type
)
}
\arguments{
\item{Transition_dataset}{TODO}

\item{Transition_name}{Character}

\item{replicatetype}{Character, (how) should replicates be generated? may be 'none',
'splitsample', 'cv' 'block-cv'.}

\item{reps}{Numeric number of replicates}

\item{balance_class}{Logical, should downsampling be implemented in RF? (TRUE/FALSE)}

\item{Downsampling_bounds}{Numeric vector, \verb{c(lower bound of class imbalance: Upper bound of class imbalance)}, if \code{balance_class = TRUE} use
to determine which datasets are modelled with downsampling}

\item{Data_period}{Character}

\item{model_folder}{Character}

\item{eval_results_folder}{Character}

\item{Model_type}{Character}
}
\description{
Wrapper function to perform model fitting, evaluation and saving with either RF or
GLM.
}
\author{
Antoine Adde adapted by Ben Black
}
