% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.grrffeatselect.r
\name{lulcc.grrffeatselect}
\alias{lulcc.grrffeatselect}
\title{lulcc.grrffeatSelect}
\usage{
lulcc.grrffeatselect(transition_result, cov_data, weight_vector, gamma)
}
\arguments{
\item{transition_result}{vector of transitions result used as response vector for GRRF}

\item{cov_data}{dataframe of predictors to be used in GRRF}

\item{weight_vector}{vector of weights to be supplied to 'classwt' argument in RF}

\item{gamma}{numeric between 0-1) controls the weight
of the normalized importance score and is called the importance coefficient.
Note that when Gamma = 0 we are essentially perform RRF (Regularized Random Forest)
i.e there is no 'Guided effect' vice versa when Gamma = 1 then we are applying
the strongest guiding effect which will lead to the most penalisation of redundant
features and hence the most concise feature sets}
}
\value{
dataframe of covariates selected by GRRF ranked according to importance (Mean decrease Gini)
}
\description{
Embedded feature selection with Guided Regularized Random Forests
}
\author{
Adapted from Deng (2013: https://arxiv.org/pdf/1306.0237.pdf)
by Antoine Adde edited by Ben Black
}
