% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.listbylulc.r
\name{lulcc.listbylulc}
\alias{lulcc.listbylulc}
\title{lulcc.listbylulc}
\usage{
lulcc.listbylulc(All_trans_list, initial_or_final)
}
\description{
Seperate list of dataframes into a nested list according to the initial LULC class of
the transition
}
\author{
Ben Black
}
