% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.modelprechecks.r
\name{lulcc.modelprechecks}
\alias{lulcc.modelprechecks}
\title{LULCC_model_pre_checks}
\usage{
lulcc.modelprechecks(config = get_config())
}
\arguments{
\item{Control_table_path}{Chr, file path of control table

date: 03-11-2022}
}
\description{
function to check that all elements required for simulation with Dinamica are
prepared
}
\author{
Ben Black
}
