% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.plotboyceandroccurves.r
\name{lulcc.plotBoyceandROCcurves}
\alias{lulcc.plotBoyceandROCcurves}
\title{lulcc.plotBoyceandROCcurves}
\usage{
lulcc.plotBoyceandROCcurves(RF_specs, GLM_specs, RF_model_num, transition_name)
}
\arguments{
\item{RF_specs}{Character specifying which group RF models to plot curves for}

\item{GLM_specs}{Character specifying which GLM models to plot curves for}

\item{RF_model_num}{Numeric to identify which RF model according to number of trees}

\item{transition_name}{Character to identify which LULC transition to produce curves for}
}
\value{
List containing seperate ROC and Boyce plots as well as a combined plot
and data for ROC plot
}
\description{
Produce Boyce and AUC curves for a specified LULC transition under specific models
}
\author{
Ben Black
}
