% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.requestfocallulcclasses.r
\name{lulcc.requestfocallulcclasses}
\alias{lulcc.requestfocallulcclasses}
\title{lulcc.finalisemodelspecifications}
\usage{
lulcc.requestfocallulcclasses(LULC_aggregation_path)
}
\arguments{
\item{LULC_aggregation_path}{Character, file path for table of LULC class aggregations}
}
\description{
Enter optimal model specification and parameters for prediction
}
\author{
Ben Black
}
