% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.savethis.r
\name{lulcc.savethis}
\alias{lulcc.savethis}
\title{Save this}
\usage{
lulcc.savethis(
  object,
  model_name = NULL,
  transition_name = NULL,
  tag = NULL,
  save_path
)
}
\arguments{
\item{object}{List, model object to be saved}

\item{model_name}{Character, model type}

\item{transition_name}{Character, transition name to be included in save path}

\item{tag}{Character, identifier for model}

\item{save_path}{Character, folder location for saving}
}
\description{
TODO text
}
\author{
Antoine Adde with edits by Ben Black
}
