% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.setparams.r
\name{lulcc.setparams}
\alias{lulcc.setparams}
\title{lulcc.setparams}
\usage{
lulcc.setparams(
  model_name,
  transition_result,
  param_grid,
  covariate_names,
  weights = 1
)
}
\arguments{
\item{model_name}{Character, abbreviation for model type accepts: "glm", "rf"}

\item{transition_result}{Vector of transition result
used as response variable in models}

\item{param_grid}{Character, file path for grid of hyperparameters}

\item{covariate_names}{Vector of covariates names to be used in model formulas}

\item{weights}{Vector of weights to be used in modelling}
}
\value{
List object of model settings
}
\description{
set hyperparameters for fitting statistical models
}
\author{
Antoine Adde (main) edited by Ben Black
}
