% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.summarisecovselection.r
\name{lulcc.summarisecovselection}
\alias{lulcc.summarisecovselection}
\title{lulcc.summarisecovselection}
\usage{
lulcc.summarisecovselection(nested_list_of_trans, split_by)
}
\arguments{
\item{nested_list_of_trans}{list of transitions with result elements
produced by lulcc.analysecovselection function}

\item{split_by}{Character determining aspect of transition
datasets to summarize by accepts:
"across_trans", "Bioregion", "Initial_lulc", "Final_lulc".}
}
\description{
Aggregate summary information on results of 2 stage covariate selection procedure
}
\author{
Ben Black
}
