% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.summarisemodelevaluation.r
\name{lulcc.summarisemodelevaluation}
\alias{lulcc.summarisemodelevaluation}
\title{lulcc.summarisemodelevaluation}
\usage{
lulcc.summarisemodelevaluation(model_eval_results, summary_metrics, plots)
}
\arguments{
\item{model_eval_results}{TODO}

\item{summary_metrics}{Vector, names of evaluation metrics to summarize by
can include: AUC, AUC.S, RMSE, Boyce, Score, threshold, Sensitivity,
Specificity, Accuracy, PPV, NPV, Jaccard, TSS Kappa, SEDI.}

\item{plots}{Logical, whether to produce summary plots (TRUE or not (FALSE))}
}
\value{
either a table summarising performance across models or
if plots = TRUE a list of the performance table and plots
}
\description{
Aggregate model evaluation metrics across transitions/bioregions
}
\author{
Ben Black
}
