% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.viabletranslist.r
\name{lulcc.viabletranslist}
\alias{lulcc.viabletranslist}
\title{lulcc.viabletranslist}
\usage{
lulcc.viabletranslist(transmatrix, inclusion_threshold)
}
\description{
Identify viable LULC transitions from a transition matrix based on a threshold
inclusion rate
}
\author{
Ben Black
}
