% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_model_finalization.r
\name{make_model_lookup_table}
\alias{make_model_lookup_table}
\title{create a df with model lookup info for each period}
\usage{
make_model_lookup_table(
  model_period,
  model_base_folder,
  trans_table,
  config = get_config()
)
}
\description{
create a df with model lookup info for each period
}
