% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_trans_potent_calc.r
\name{model_neigh_preds}
\alias{model_neigh_preds}
\title{E.2- Dynamic predictors: Neighbourhood predictors}
\usage{
model_neigh_preds(current_LULC_raster, config, period_tag, lulc_rat)
}
\description{
E.2- Dynamic predictors: Neighbourhood predictors
}
