% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.multi_input_class.r
\docType{class}
\name{multi.input-class}
\alias{multi.input-class}
\alias{pipe.multi.input}
\title{define multi.input class}
\description{
An S4 class to store evaluation data
}
\section{Slots}{

\describe{
\item{\code{mod}}{Name of model algorithm to be called (character) e.g., 'glm'}

\item{\code{args}}{A list of arguments to be supplied to the model fitting alrogrithm}

\item{\code{tag}}{The name of the model setup (character)}

\item{\code{weight}}{Should observations be weighted based on their prevalence?}
}}

\author{
Philipp Brun (main) with edits by Antoine Adde
}
