% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_predictor_prep.r
\name{nhood_predictor_prep}
\alias{nhood_predictor_prep}
\title{Nhood_data_prep: Neighbourhood effect predictor layer preparation}
\usage{
nhood_predictor_prep(config = get_config(), redo_random_matrices = FALSE)
}
\arguments{
\item{config}{list of configuration parameters}

\item{redo_random_matrices}{logical indicating whether to re-generate the random matrices}
}
\description{
When devising neighborhood effect predictors there are two considerations:
1.The size of the neighborhood (no. of cells: n)
2.The decay rate from the central value outwards and to a lesser extent the choice of
method for interpolating the decay rate values.
}
\details{
This script will follow the process for automatic rule detection procedure (ARD)
devised by Roodposhti et al. (2020) to test various permutations of these values; for
more details refer to:
http://www.spatialproblems.com/wp-content/uploads/2019/09/ARD.html
}
\author{
Ben Black
}
