% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.multi.r
\name{pipe.multi}
\alias{pipe.multi}
\title{define multi function}
\usage{
pipe.multi(mod, args, tag = "", step = FALSE, weight = FALSE)
}
\arguments{
\item{mod}{a character with the name of the function to be called. E.g. "gam"}

\item{args}{a list with arguments to be passed to the function specified in mod}

\item{tag}{character with name for model set-up}
}
\value{
Object of class 'multi'
}
\description{
Define settings for function that should be supplied to wsl.flex;
Create a multi.input object that efficiently stores model specifications
}
\author{
Philipp Brun (main) with edits by Antoine Adde
}
