% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.summary_wsl_evaluation.r
\name{pipe.summary}
\alias{pipe.summary}
\title{define summary function for wsl.evaluation objects}
\usage{
pipe.summary(object)
}
\description{
TODO text
}
\author{
Philipp Brun (main) with edits by Antoine Adde
}
