% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_prep.r
\name{region_prep}
\alias{region_prep}
\title{Region Preparation: region_prep}
\usage{
region_prep(config = get_config())
}
\description{
Preparing Rasters of regions in Peru based upon the ecoregions classification from 2017 to sub-divide data for regional
modelling 
}
\author{
Ben Black
}
