% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.all_metrics_function.r
\name{sens}
\alias{sens}
\title{Evaluation function:}
\usage{
sens(a, c)
}
\description{
Small functions to calculate common threshold based confusion matrix indices
}
\author{
Philipp Brun (main) with edits by Antoine Adde
}
