% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_predictor_prep.r
\name{simulation_predictor_prep}
\alias{simulation_predictor_prep}
\title{Simulation Predictor Preparation}
\usage{
simulation_predictor_prep(config = get_config())
}
\arguments{
\item{config}{List containing configuration parameters, defaults to result of
\code{\link[=get_config]{get_config()}}}
}
\value{
None. Files are written to disk at locations specified in config.
}
\description{
Prepares predictor data lookup tables for simulation time steps. This includes:
\itemize{
\item Processing future economic scenarios for employment data
\item Preparing mechanisms for future population projections
\item Processing future climatic data
\item Creating predictor lookup tables for simulation periods
}
}
\author{
Ben Black
}
