% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_trans_tables_prep.r
\name{simulation_trans_tables_prep}
\alias{simulation_trans_tables_prep}
\title{Prepare transition tables for land use and land cover (LULC) simulation}
\usage{
simulation_trans_tables_prep(config = get_config())
}
\arguments{
\item{config}{Configuration list containing paths and parameters. If not provided,
it will be obtained from get_config()}
}
\value{
No return value. The function creates and saves various transition tables
and auxiliary data for LULC simulation.
}
\description{
Create tables of transition rates for future time points for use in LULCC allocation
within Dinamica EGO. Read historical LULC data, create linearly extrapolated
transition rates, and modify them according to different scenarios. The process
involves several steps:
}
\section{Overall process}{

\enumerate{
\item Prepare data and folders
\item Create folder structure for scenario-specific transition rate tables
\item Interpolate/extrapolate transition rates into future years
\item Create transition tables for calibration periods
\item Calculate expected LULC class areal change
\item Modify transition rates for different scenarios
\item Check the resulting transition areas against desired LULC class areas
\item Calculate final modified transition rates
}
}

\author{
Ben Black
}
