% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interventions_prep.r
\name{spatial_interventions_prep}
\alias{spatial_interventions_prep}
\title{Spatial_prob_perturb}
\usage{
spatial_interventions_prep(config = get_config())
}
\arguments{
\item{config}{list of config options}
}
\description{
Prepare spatial layers for perturbation of cellular transition probabilities in
simulation steps
}
\author{
Ben Black
}
