% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_model_finalization.r
\name{trans_model_finalization}
\alias{trans_model_finalization}
\title{Trans_model_finalization}
\usage{
trans_model_finalization(config = get_config())
}
\arguments{
\item{config}{A list of configuration parameters}
}
\description{
Re-fit LULCC transition models using optimal specifcation after calling
\code{\link[=lulcc.finalisemodelspecifications]{lulcc.finalisemodelspecifications()}}
}
\author{
Ben Black
}
