% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_dataset_prep.r
\name{transition_dataset_prep}
\alias{transition_dataset_prep}
\title{transition_dataset_creation}
\usage{
transition_dataset_prep(config = get_config())
}
\arguments{
\item{config}{list of configuration parameters}
}
\description{
Script for gathering the layers of LULC (dependent variable) and predictors for each
historic period then separating into viable LULC transitions at the scale of
Switzerland and Bioregions
}
\author{
Ben Black
}
