% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_feature_selection.r
\name{transition_feature_selection}
\alias{transition_feature_selection}
\title{Feature_selection}
\usage{
transition_feature_selection(config = get_config())
}
\description{
Performing collinearity based and embedded feature selection with Guided
Regularized Random Forests
}
\author{
: Ben Black
}
