% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_identification.r
\name{transition_identification}
\alias{transition_identification}
\title{Transition identification}
\usage{
transition_identification(config = get_config())
}
\description{
Transition_identification: Using land use data from calibration (historic) periods to identify
LULC transitions to be modeled in future simulations
}
\author{
Ben Black
}
