% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{util}
\alias{util}
\alias{ensure_dir}
\title{Utility functions for evoland}
\usage{
ensure_dir(dir)
}
\description{
Sundry functions that are often used
}
\section{Functions}{
\itemize{
\item \code{ensure_dir()}: Ensure that a directory exists

}}
