% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_predictor_prep.r
\name{calibration_predictor_prep}
\alias{calibration_predictor_prep}
\title{Preparation of Predictor Data}
\usage{
calibration_predictor_prep(config)
}
\description{
Not yet clear why this script was originally called \code{calibration_predictor_prep.r} -
100m resolution, crs and extent prepared layers are saved seperately in a
Use predictor table to prepare predictor layers at a uniform
100m resolution, crs and extent prepared layers are saved seperately in a
and then combined into raster stacks for easy loading
Date: 01-08-2021
Author: Ben Black
}
