% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_trans_potent_calc.r
\name{dinamica_trans_potent_calc}
\alias{dinamica_trans_potent_calc}
\title{dinamica_trans_potent_calc}
\usage{
dinamica_trans_potent_calc(
  config = get_config(),
  simulation_num = integer(),
  time_step = integer()
)
}
\arguments{
\item{config}{A list containing configuration parameters and file paths. Default:
result of \code{get_config()}.}

\item{simulation_num}{Integer. Simulation ID number to select simulation parameters.}

\item{time_step}{Integer. The current time step for which transition potentials are
calculated.}
}
\description{
Calculates the transition potential for land use/land cover (LULC) changes using
fitted statistical models. Supports both calibration and simulation modes, and can
optionally apply scenario-specific spatial interventions. Generates dynamic
predictors (e.g., population, neighbourhood), predicts transition probabilities for
each possible LULC transition, rescales probabilities, applies interventions if
required, and saves the resulting probability maps as raster files.
}
\details{
\enumerate{
\item Loads simulation parameters and configuration.
\item Loads current LULC raster and layerizes it by class.
\item Loads suitability and accessibility predictors, adapting to calibration or
simulation mode.
\item In simulation mode, generates dynamic predictors such as municipal population and
neighbourhood statistics.
\item Stacks all predictors and converts them to a dataframe for model prediction.
\item Loads transition models and predicts transition probabilities for each transition
and region.
\item Rescales probabilities to ensure they sum to 1 for each cell.
\item Optionally applies scenario-specific spatial interventions to the probabilities.
\item Saves probability maps for each transition as raster files.
}
}
\seealso{
\code{\link[=lulcc.spatprobmanipulation]{lulcc.spatprobmanipulation()}}
}
\author{
Ben Black, Carlson Büth
}
