% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_utils.r
\name{dinamica_utils}
\alias{dinamica_utils}
\alias{exec_dinamica}
\alias{run_dinamica_extrapolation}
\alias{process_dinamica_script}
\title{Dinamica Utility Functions}
\usage{
exec_dinamica(
  model_path,
  disable_parallel = TRUE,
  log_level = NULL,
  additional_args = NULL,
  verbose = FALSE
)

run_dinamica_extrapolation(
  run_modelprechecks = TRUE,
  config = get_config(),
  work_dir = format(Sys.time(), "\%Y-\%m-\%d_\%Hh\%Mm\%Ss"),
  verbose = FALSE
)

process_dinamica_script(infile, outfile, mode = "encode", check = TRUE)
}
\arguments{
\item{model_path}{Path to the .ego model file to run. Any submodels must be included
in a directory of the exact form \verb{basename(modelpath)_ego_Submodels}, \href{https://csr.ufmg.br/dinamica/dokuwiki/doku.php?id=submodels}{see wiki}}

\item{disable_parallel}{Whether to disable parallel steps (default TRUE)}

\item{log_level}{Logging level (1-7, default NULL)}

\item{additional_args}{Additional arguments to pass to DinamicaConsole, see
\code{DinamicaConsole -help}}

\item{verbose}{bool}

\item{run_modelprechecks}{bool, Validate that everything's in place for a model run.
Set to false for calibration runs.}

\item{work_dir}{Working dir, where to place ego files and control table}

\item{infile}{Input file path. Treated as input if passed AsIs using \code{base::I()}}

\item{outfile}{Output file path (optional)}

\item{mode}{Character, either "encode" or "decode"}

\item{check}{Default TRUE, simple check to ensure that you're handling what you're expecting}

\item{simctrl_tbl_template}{Path to the template simulation control table}
}
\description{
Functions to interact with Dinamica from R
}
\section{Functions}{
\itemize{
\item \code{exec_dinamica()}: Execute a Dinamica .ego file using \code{DinamicaConsole}

\item \code{run_dinamica_extrapolation()}: Run a Dinamica EGO extrapolation simulation

\item \code{process_dinamica_script()}: Encode or decode raw R and Python code chunks in .ego
files and their submodels to/from base64

}}
