% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.covfilter.r
\name{lulcc.covfilter}
\alias{lulcc.covfilter}
\title{lulcc.covfilter}
\usage{
lulcc.covfilter(cov_data, trans_result, method, weights, corcut = 0)
}
\arguments{
\item{cov_data}{TODO}

\item{trans_result}{TODO}

\item{method}{a character of the method used to rank covariates
accepts: "IQR", "IQR.M", "COR.P", "COR.S", "WIL", "GLM", "MAD.M"}

\item{weights}{vector of 'prior weights' to be used in the glm fitting
process. Should be NULL or a numeric vector.}

\item{corcut}{numeric value for correlation cut-off to remove covariates}
}
\value{
a dataframe of the collinearity filtered covariates
}
\description{
Covariate collinearity testing
}
\author{
Antoine Adde (main) edited by Ben Black
}
