% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.downloadunzip.r
\name{lulcc.downloadunzip}
\alias{lulcc.downloadunzip}
\title{lulcc.downloadunzip}
\usage{
lulcc.downloadunzip(url, save_dir, filename = basename(url))
}
\arguments{
\item{url}{character string url to download folder from
produced by function lulcc.evalfeatureselection}

\item{filename}{defaults to the \code{basename(url)}}

\item{save_path}{character string path of directory to save unzipped files in}
}
\description{
Download folder from URL and unzip to target directory
}
\author{
Ben Black
}
