% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.testrastercompatibility.r
\name{lulcc.TestRasterCompatibility}
\alias{lulcc.TestRasterCompatibility}
\title{Confirm Raster compatibility prior to stacking}
\usage{
lulcc.TestRasterCompatibility(Rasterlist, ExemplarRaster)
}
\description{
TODO text
}
\author{
Ben Black
}
